prime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -5,8 +5,8 @@
 */
 int main(void)
 {
-	unsigned long  int n = 612852475143;
-	unsigned long int div = 2, ans = 0, maxFact;
+	int n = 612852475143;
+	int div = 2, ans = 0, maxFact;
 
 	while (n != 0)
 	{
